requirements.txt
streamlit
python-whois
requests
